refactor: relocate `svg` folder for better organization
Moved `svg` folder from `components/ui/svg` to `components/svg` to improve project structure and make SVG assets easier to locate and manage.
