+++ src/components/svg/ZekeLogoIcon.jsx
